refactor: restructure vault editing components and styles, including EditVaultPage, VaultBackupPage, VaultDeletePage, and VaultRenamePage for improved organization and maintainability
