+++ clients/desktop/src/navigation/router.tsx
@@ -11,11 +11,6 @@ import { LauncherObserver } from '../launcher/components/LauncherObserver'
 import { ManageDklsPage } from '../mpc/dkls/ManageDklsPage'
 import { OnboardingPage } from '../onboarding/components/OnboardingPage'
 import { IncompleteOnboardingOnly } from '../onboarding/IncompleteOnboardingOnly'
-import EditVaultPage from '../pages/editVault/EditVaultPage'
-import { VaultBackupPage } from '../pages/editVault/vaultBackupSettings/VaultBackupPage'
-import DeleteVaultPage from '../pages/editVault/vaultDeleteSettings/DeleteVaultPage'
-import VaultDetailsPage from '../pages/editVault/vaultDetailsSettings/VaultDetailsPage'
-import VaultRenamePage from '../pages/editVault/vaultRenameSettings/VaultRenamePage'
 import RegisterForAirdropPage from '../pages/registerForAirdrop/RegisterForAirdropPage'
 import SettingsVaultPage from '../pages/vaultSettings/SettingsVaultPage'
 import AddressBookSettingsPage from '../pages/vaultSettings/vaultAddressBook/AddressBookSettingsPage'
@@ -32,6 +27,11 @@ import { ActiveVaultGuard } from '../vault/components/ActiveVaultGuard'
 import { EmptyVaultsOnly } from '../vault/components/EmptyVaultsOnly'
 import { VaultPage } from '../vault/components/VaultPage'
 import { DepositPage } from '../vault/deposit/DepositPage'
+import EditVaultPage from '../vault/edit/EditVaultPage'
+import { VaultBackupPage } from '../vault/edit/vaultBackupSettings/VaultBackupPage'
+import DeleteVaultPage from '../vault/edit/vaultDeleteSettings/DeleteVaultPage'
+import VaultDetailsPage from '../vault/edit/vaultDetailsSettings/VaultDetailsPage'
+import VaultRenamePage from '../vault/edit/vaultRenameSettings/VaultRenamePage'
 import { ImportVaultFromFilePage } from '../vault/import/components/ImportVaultFromFilePage'
 import { ImportVaultPage } from '../vault/import/components/ImportVaultPage'
 import { JoinKeygenPage } from '../vault/keygen/join/JoinKeygenPage'
