+++ clients/desktop/src/vault/edit/EditVaultPage.styles.tsx
