+++ clients/desktop/src/vault/edit/vaultDeleteSettings/DeleteVaultPage.styles.tsx
