+++ clients/desktop/src/vault/edit/vaultDeleteSettings/DeleteVaultPage.tsx
