+++ clients/desktop/src/vault/edit/vaultDetailsSettings/VaultDetailsPage.styles.tsx
