+++ clients/desktop/src/vault/edit/vaultDetailsSettings/VaultDetailsPage.tsx
