+++ clients/desktop/src/vault/edit/vaultRenameSettings/VaultRenamePage.styles.tsx
